refactor: cleaning up types and naming
