+++ runtime/lib/HTTPClient/function-lambda/exceptions/function-code-not-found.exception.ts
@@ -1,0 +1,7 @@
+import { LambdaException } from './lambda.exception';
+
+export class FunctionCodeNotFoundException extends LambdaException {
+  get message(): string {
+    return `Unable to retrieve Function code in lambda`;
+  }
+}
