+++ runtime/lib/HTTPClient/function-lambda/exceptions/invalid-runtime-command.exception.ts
