+++ runtime/lib/HTTPClient/function-lambda/exceptions/lambda.exception.ts
