+++ runtime/lib/HTTPClient/function-lambda/exceptions/runtime-error.exception.ts
