--- runtime/lib/HTTPClient/function-lambda/interface.ts
@@ -1,14 +1,0 @@
-interface BaseFunctionLambdaRequest {
-  variables: Record<string, unknown>;
-  enableLog?: boolean;
-}
-
-interface FunctionLambdaRequestWithReference extends BaseFunctionLambdaRequest {
-  codeId: string;
-}
-
-interface FunctionLambdaRequestWithCode extends BaseFunctionLambdaRequest {
-  code: string;
-}
-
-export type FunctionLambdaRequest = FunctionLambdaRequestWithReference | FunctionLambdaRequestWithCode;
