--- runtime/lib/Handlers/function/lib/execute-lambda/lambda-error-code.enum.ts
@@ -1,13 +1,0 @@
-import { Enum } from '@voiceflow/dtos/build/cjs/utils/type/enum.util';
-
-export const LambdaErrorCode = {
-  General: 'general:error',
-
-  InvalidRequest: 'handler:invalid-request',
-
-  SandboxRuntimeError: 'sandbox:runtime-error',
-
-  SandboxModuleResolution: 'sandbox:module-resolution',
-} as const;
-
-export type LambdaErrorCode = Enum<typeof LambdaErrorCode>;
